Distinguish between numbers with and without fractional component
Currently, parsing the number "1" and writing it out again yields "1.0".
This asymetry can be troublesome when sending messages to software
using other parsers, as they might reject "1.0" as a valid integer.

This PR distinguishes between numbers that contain a fractional (or
exponential) component and those that don't.

Used the [JSON spec](https://www.rfc-editor.org/rfc/rfc8259#section-6)
for reference and inspiration.