Round-trips for integers introduce fractional part
Currently, parsing the number "1" and writing it out again yields "1.0".
This asymetry can be troublesome when sending messages to software
that distinguishes between integer and non-integer numbers, as those
might reject "1.0" as a valid integer.

This first commit introduces a test that demonstrates the issue. A subsequent
commit will implement a solution for it.